+++ src/FluentValidationServiceProvider.php
@@ -2,17 +2,39 @@
 
 namespace Saritasa\Laravel\Validation;
 
-use Illuminate\Support\ServiceProvider;
-use Illuminate\Contracts\Validation\Factory as IValidatorFactory;
+use Illuminate\Validation\ValidationServiceProvider;
+use Propaganistas\LaravelPhone\PhoneServiceProvider;
 
 /**
  * Service provider substitutes default Laravel's validation factory
  * @see FluentValidatorFactory
  */
-class FluentValidationServiceProvider extends ServiceProvider
+class FluentValidationServiceProvider extends ValidationServiceProvider
 {
-    public function boot()
+    protected function registerValidationFactory()
     {
-        $this->app->bind(IValidatorFactory::class, FluentValidatorFactory::class);
+        $this->app->singleton('validator', function ($app) {
+            $validator = new FluentValidatorFactory($app['translator'], $app);
+
+            // The validation presence verifier is responsible for determining the existence of
+            // values in a given data collection which is typically a relational database or
+            // other persistent data stores. It is used to check for "uniqueness" as well.
+            if (isset($app['db'], $app['validation.presence'])) {
+                $validator->setPresenceVerifier($app['validation.presence']);
+            }
+
+            return $validator;
+        });
+    }
+
+    /**
+     * Register any application services.
+     *
+     * @return void
+     */
+    public function register()
+    {
+        parent::register();
+        $this->app->register(PhoneServiceProvider::class);
     }
 }